requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
discord.py==1.2.4
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
multidict==4.5.2
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.6
websockets==6.0
xlwt==1.3.0
yarl==1.3.0
